+++ packages/components/FocusZone/src/FocusZone.types.ts
@@ -18,23 +18,40 @@ export interface FocusZoneProps {
   defaultTabbableElement?: React.RefObject<React.Component>;
 
   /**
-   ** Defines which arrows to react to
+   ** Defines which arrow keys to react to
    */
   focusZoneDirection?: FocusZoneDirection;
 
   /**
-   ** If set, the FocusZone will not be tabbable and keyboard navigation will be disabled
+   * Disables the FocusZone, with slightly different behavior on macOS and win32.
+   * On win32, the FocusZone will not be tabbable and keyboard navigation will be disabled.
+   * on macOS, the FocusZone will "pass through" events, and the children will respond as if there was no FocusZone.
    */
   disabled?: boolean;
 
-  /* Circular Navigation prop */
+  /**
+   * Circular Navigation prop
+   * @platform win32
+   */
   isCircularNavigation?: boolean;
 
   /**
-   *Allows for 2D navigation. This navigation strategy takes into account the position of elements
-   on screen, and navigates in the direction the user selects to the nearest element.
+   * Allows for 2D navigation. This navigation strategy takes into account the position of elements
+   * on screen, and navigates in the direction the user selects to the nearest element.
+   * @platform win32
    */
   use2DNavigation?: boolean;
+
+  /**
+   * Moves focus between all focusable views, rather than just key views.
+   *
+   * On macOS, not every focusable view is a key view (i.e: we can press Tab to move focus to it).
+   * Rather, there is a system preference to toggle which views are in the key view loop.
+   * This prop allows you to focus on all focusable views, rather than just key views.
+   * For more info, see https://microsoft.github.io/apple-ux-guide/KeyboardFocus.html
+   * @platform macOS
+   */
+  forceFocusMacOS?: boolean;
 
   /**
    * Callback called when “focus” event triggered in FocusZone
@@ -42,8 +59,14 @@ export interface FocusZoneProps {
   onFocus?: (e?: any) => void;
 }
 
-export interface NativeProps extends Omit<FocusZoneProps, 'isCircularNavigation'> {
-  navigateAtEnd?: NavigateAtEnd;
+// Props on JS FocusZone that don't exist in the native module
+interface NonNativeProps {
+  isCircularNavigation: FocusZoneProps['isCircularNavigation'];
+  forceFocusMacOS: FocusZoneProps['forceFocusMacOS'];
+}
+export interface NativeProps extends Exclude<FocusZoneProps, NonNativeProps> {
+  navigateAtEnd?: NavigateAtEnd; // win32 only
+  forceFocus?: boolean; // macOS only
 }
 
 export type FocusZoneDirection =